requirements.txt
#Usage: pip install -r requirements.txt

#Flask Framework
Flask==3.0.2

Flask-Cors==4.0.0
Flask-PyMongo==2.3.0

bcrypt==4.1.2
blurhash-python==1.2.2
boto3==1.34.99
PyJWT==2.8.0
pymongo==4.7.1
Werkzeug==3.0.2
python-dotenv==1.0.1
gunicorn==22.0.0
